Fixed issues, including: renaming height checker moved checking z error to height holder removed redundant variables removed locks where they were not needed misc fixes
