+++ param/motion_planner.yaml
@@ -44,4 +44,4 @@ k_term_tracking_y: .35
 # p-term for holding Z height
 k_term_hold_z: 1.25
 #maximum z error for heigth holder
-max_z_error: .5
+max_z_error: .7
